requirements-dev.txt
django-silk>=3.0.0
django-webpack-loader
